Move widget traits and types to kas::core module
