+++ src/core/widget.rs
